Use explicit bit-count types.
When you compile the code with sanity checks, the warnings are emitted in the console, stating that a buffer overrun will happen if the respective functions with some magic numbers are called.

The problem is that the header in question seems ANCIENT. Somewhere in the distant past, the longs could really be 64 bits, but almost none use that convention currently, as longs are 32 bits under most compilers.

I have just altered the default types in the header to explicitly state the sizes of the ints. This way, there are no more warnings emitted regarding that issue.